+++ R/step-subset-select.R
@@ -25,6 +25,7 @@ select.dtplyr_step <- function(.data, ...) {
   if (length(vars) == 0) {
     j <- 0L
     groups <- .data$groups
+    is_unnamed <- TRUE
   } else {
     groups <- rename_groups(.data$groups, vars)
     vars <- simplify_names(vars)
@@ -36,7 +37,18 @@ select.dtplyr_step <- function(.data, ...) {
     j <- call2(".", !!!syms(vars))
   }
 
-  out <- step_subset_j(.data, vars = names(locs), groups = character(), j = j)
+  if (is_copied(.data) && is_unnamed && !can_merge_subset(.data)) {
+    # Drop columns by reference if:
+    #  * Data has been copied (implicitly or explicitly)
+    #  * There is no renaming in the select statement
+    #  * The selection can't be combined with a prior `i` step. Ex: dt[x < 7, .(x, y)]
+    vars_drop <- setdiff(.data$vars, vars)
+    out <- remove_vars(.data, vars_drop)
+    out <- step_colorder(out, vars)
+  } else {
+    out <- step_subset_j(.data, vars = names(locs), groups = character(), j = j)
+  }
+
   step_group(out, groups)
 }
 
@@ -94,7 +106,6 @@ ensure_group_vars <- function(loc, names, groups) {
   loc
 }
 
-
 rename_groups <- function(groups, vars) {
   old2new <- set_names(names(vars), vars)
   groups[groups %in% names(old2new)] <- old2new[groups]
@@ -104,4 +115,15 @@ rename_groups <- function(groups, vars) {
 simplify_names <- function(vars) {
   names(vars)[vars == names(vars)] <- ""
   vars
+}
+
+remove_vars <- function(.data, vars) {
+  if (is_empty(vars)) {
+    return(.data)
+  }
+  out <- step_subset(
+    .data, groups = character(), j = expr(!!unique(vars) := NULL),
+    vars = setdiff(.data$vars, vars)
+  )
+  group_by(out, !!!syms(.data$groups))
 }