+++ R/step-subset-transmute.R
@@ -46,7 +46,7 @@ transmute.dtplyr_step <- function(.data, ...) {
   vars <- union(group_vars(.data), names(dots))
   out <- step_subset_j(.data, vars = vars, j = j)
   if (dots_list$need_removal_step) {
-    out <- remove_vars(out, dots_list$vars_removed)
+    out <- select(out, -tidyselect::all_of(dots_list$vars_removed))
   }
 
   out