+++ NEWS.md
@@ -39,6 +39,7 @@
 
 * `count()` properly handles grouping variables (#356)
 
+* `select()` now drops columns by reference when possible for improved performance (#367)
 * `arrange()` now utilizes `setorder()` when possible for improved performance (#364)
 
 * `slice()` uses an intermediate variable to reduce computation time of row selection (#377)